migrate to nodejs to support cloud function on Rpi 4 (ARM)
