+++ appwrite_functions/averageHourlyPing/.env
@@ -1,0 +1,10 @@
+# this file is there to simulate the variable available from appwrite
+
+#APPWRITE_API_ENDPOINT=https://appwrite.DOMAINE.COM/v1
+#APPWRITE_PROJECT_ID=YYYY
+#APPWRITE_API_KEY=XXXX
+
+#COLLECTION_ID_HOURLYPINGS=1234
+#COLLECTION_ID_DAILYPINGS=1235
+#COLLECTION_ID_PROJECTS=1236
+#COLLECTION_ID_PINGS=1237
