--- appwrite_functions/averageHourlyPing/.vscode/settings.json
@@ -1,8 +1,0 @@
-{
-  "deno.enable": true,
-  "deno.lint": true,
-  "deno.unstable": true,
-  "deno.suggest.imports.hosts": {
-    "https://deno.land": true
-  }
-}
