+++ appwrite_functions/pingCronWrapper/.env
@@ -1,0 +1,8 @@
+# this file is there to simulate the variable available from appwrite
+
+#APPWRITE_API_ENDPOINT=https://appwrite.DOMAINE.COM/v1
+#APPWRITE_PROJECT_ID=YYYY
+#APPWRITE_API_KEY=XXXX
+
+#FUNCTION_ID_PINGSERVER=1234
+#COLLECTION_ID_PROJECTS=1235
