+++ appwrite_functions/pingCronWrapper/package.json
@@ -1,0 +1,15 @@
+{
+    "name": "pingcronwrapper",
+    "version": "1.0.0",
+    "description": "",
+    "main": "pingCronWrapper.js",
+    "scripts": {
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+        "dotenv": "^10.0.0",
+        "node-appwrite": "^2.4.0"
+    }
+}
