+++ appwrite_functions/pingServer/.env
@@ -1,0 +1,10 @@
+# this file is there to simulate the variable available from appwrite
+
+#APPWRITE_API_ENDPOINT=https://appwrite.DOMAINE.COM/v1
+#APPWRITE_PROJECT_ID=YYYY
+#APPWRITE_API_KEY=XXXX
+
+#APPWRITE_FUNCTION_DATA=
+#COLLECTION_ID_PINGS=1234
+#COLLECTION_ID_PROJECTS=1235
+#SLOW_RESPONSE_TRESHOLD=1236
