+++ appwrite_functions/pingServer/package.json
@@ -1,0 +1,16 @@
+{
+    "name": "pingserver",
+    "version": "1.0.0",
+    "description": "",
+    "main": "pingServer.js",
+    "scripts": {
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+        "dotenv": "^10.0.0",
+        "node-appwrite": "^2.4.0",
+        "node-fetch": "^2.6.1"
+    }
+}
