+++ appwrite_setup/setup/functions.json
@@ -9,7 +9,7 @@
       "dateCreated": 1628684927,
       "dateUpdated": 1628684961,
       "status": "disabled",
-      "runtime": "deno-1.8",
+      "runtime": "node-15.5",
       "tag": "",
       "vars": [],
       "events": [],
@@ -28,7 +28,7 @@
       "dateCreated": 1628684927,
       "dateUpdated": 1628684961,
       "status": "disabled",
-      "runtime": "deno-1.8",
+      "runtime": "node-15.5",
       "tag": "",
       "vars": [],
       "events": [],
@@ -47,7 +47,7 @@
       "dateCreated": 1628684992,
       "dateUpdated": 1628686246,
       "status": "disabled",
-      "runtime": "deno-1.8",
+      "runtime": "node-15.5",
       "tag": "",
       "vars": [],
       "events": [],
